Backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
boto3==1.34.59
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiosqlite==0.20.0
python-multipart==0.0.9
cors==1.0.1
bcrypt==4.1.2
email-validator==2.1.0.post1
reportlab==4.2.2
scikit-learn==1.4.1.post1
pandas==2.2.1
prophet==1.1.5
beautifulsoup4==4.12.3
requests==2.31.0
